requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
asyncpg==0.29.0
earthengine-api==0.1.382
google-genai
httpx>=0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
web3>=6.0.0
eth-account>=0.9.0
py-solc-x>=1.12.0
requests>=2.28.0